refactor(worker): extract main loop into internal/worker
Move all non-main logic from cmd/worker into a new internal/worker
package so upcoming e2e tests can call worker.Run(ctx, store, cfg)
in-process. Behavior is unchanged; env reads in the request path
(WORKSPACE_ROOT, AIOPS_MIRROR_ROOT, GITEA_BASE_URL, GITEA_TOKEN) are
replaced by cfg.* fields threaded through Config. Unexported helpers
that external tests (now package worker_test) require are exported:
Emit, ErrSummary, SummarizeVerifyResults, BuildPRBody, PRBodySummaryCap,
AppendRunSummaryDirective, RunRunnerWithTimeout, CreatePRWith,
ResolveWorkflow, RunVerifyPhase, EventEmitter, PRClient, RunTaskError.
