+++ internal/worker/print_config.go
@@ -1,4 +1,4 @@
-package main
+package worker
 
 import (
 	"encoding/json"
