+++ internal/worker/print_config_test.go
@@ -1,4 +1,4 @@
-package main
+package worker
 
 import (
 	"bytes"
